Read Receipt for chat messages
This commit introduces read receipts for chat messages. This is accomplished by sending an
acknowledgment message that contains the crc32 of the recieved message.

Chat messages now contain a "[read]" or "[unread<crc32>]" next to their timestamp.
Upon sending a chat message the line below the message will display the unread status.
After a message is read, if read receipt is enabled, the client will send an ack message
that when parsed by the original sender will update the display to say the read status.